requirements.txt
python-hglib >= 1.6
pyyaml
